- sonar fixes
